refactor: toggle header component → ToggleWithHeader.swift
